mailsender: file is created for sendgridmail
