+++ go/src/socialapi/workers/email/sender/sender.go
@@ -2,11 +2,8 @@
 package sender
 
 import (
-	"fmt"
-
 	"github.com/koding/bongo"
 	"github.com/koding/logging"
-	"github.com/sendgrid/sendgrid-go"
 	"github.com/streadway/amqp"
 )
 
@@ -19,11 +16,6 @@ type Emailer interface {
 type Controller struct {
 	log     logging.Logger
 	emailer Emailer
-}
-
-// SendGridMail includes the required Sendgrid struct
-type SendGridMail struct {
-	Sendgrid *sendgrid.SGClient
 }
 
 // New Creates a new controller for mail worker
@@ -54,37 +46,4 @@ func (c *Controller) DefaultErrHandler(delivery amqp.Delivery, err error) bool {
 	delivery.Nack(false, true)
 
 	return false
-}
-
-// Send implements Emailer interface
-func (sg *SendGridMail) Send(m *Mail) error {
-	message := sendgrid.NewMail()
-
-	if err := message.AddTo(m.To); err != nil {
-		return err
-	}
-
-	if err := message.SetFrom("mail@koding.com"); err != nil {
-		return err
-	}
-
-	if m.From != "" {
-		message.SetFrom(m.From)
-	}
-
-	message.SetText(m.Text)
-	message.SetHTML(m.HTML)
-	message.SetSubject(m.Subject)
-	message.SetFromName(m.FromName)
-	if m.ReplyTo != "" {
-		if err := message.SetReplyTo(m.ReplyTo); err != nil {
-			return err
-		}
-	}
-
-	if err := sg.Sendgrid.Send(message); err != nil {
-		return fmt.Errorf("an error occurred while sending email error as %+v ", err.Error())
-	}
-
-	return nil
 }
